Pull out logic for sending the responses into its own file
